[#101] Rename: move existing Modal files to modals folders
